[API] Mv neutralize-challenge to certif controller
