+++ api/tests/unit/application/certifications/certification-controller_test.js
@@ -4,6 +4,8 @@ const certificationController = require('../../../../lib/application/certificati
 const usecases = require('../../../../lib/domain/usecases');
 const certificationSerializer = require('../../../../lib/infrastructure/serializers/jsonapi/certification-serializer');
 const certificationAttestationPdf = require('../../../../lib/infrastructure/utils/pdf/certification-attestation-pdf');
+const events = require('../../../../lib/domain/events');
+const ChallengeNeutralized = require('../../../../lib/domain/events/ChallengeNeutralized');
 
 describe('Unit | Controller | certifications-controller', () => {
 
@@ -123,6 +125,39 @@ describe('Unit | Controller | certifications-controller', () => {
       });
       expect(response.source).to.deep.equal(attestationPDF);
       expect(response.headers['Content-Disposition']).to.contains('attachment; filename=attestation-pix-20181003.pdf');
+    });
+  });
+
+  describe('#neutralizeChallenge', () => {
+    it('neutralizes the challenge and dispatches the event', async () => {
+      // given
+      const request = {
+        payload: {
+          data: {
+            attributes: {
+              certificationCourseId: 1,
+              challengeRecId: 'rec43mpMIR5dUzdjh',
+            },
+          },
+        },
+        auth: { credentials: { userId: 7 } },
+      };
+      const eventToBeDispatched = new ChallengeNeutralized({ certificationCourseId: 1, juryId: 7 });
+      sinon.stub(usecases, 'neutralizeChallenge').withArgs({
+        certificationCourseId: 1,
+        challengeRecId: 'rec43mpMIR5dUzdjh',
+        juryId: 7,
+      }).resolves(eventToBeDispatched);
+      sinon.stub(events, 'eventDispatcher').value({
+        dispatch: sinon.stub(),
+      });
+
+      // when
+      const response = await certificationController.neutralizeChallenge(request, hFake);
+
+      // then
+      expect(events.eventDispatcher.dispatch).to.have.been.calledWithExactly(eventToBeDispatched);
+      expect(response.statusCode).to.equal(204);
     });
   });
 });
