+++ api/tests/unit/application/certifications/index_test.js
@@ -1,16 +1,16 @@
 const { sinon, expect, HttpTestServer } = require('../../../test-helper');
 const securityPreHandlers = require('../../../../lib/application/security-pre-handlers');
 
-const assessmentResultSessionController = require('../../../../lib/application/assessment-results/assessment-result-controller');
-const moduleUnderTest = require('../../../../lib/application/assessment-results');
+const certificationController = require('../../../../lib/application/certifications/certification-controller');
+const moduleUnderTest = require('../../../../lib/application/certifications');
 
-describe('Unit | Application | Assessment-Results | Routes', () => {
+describe('Unit | Application | Certification | Routes', () => {
 
-  context('POST /api/assessment-results/neutralize-challenge', () => {
+  context('POST /api/admin/certification/neutralize-challenge', () => {
 
     it('rejects access if the logged user is not a Pix Master', async () => {
       // given
-      sinon.stub(assessmentResultSessionController, 'neutralizeChallenge').returns('ok');
+      sinon.stub(certificationController, 'neutralizeChallenge').returns('ok');
       sinon.stub(securityPreHandlers, 'checkUserHasRolePixMaster');
       securityPreHandlers.checkUserHasRolePixMaster.callsFake((request, h) => h.response().code(403).takeover());
       const httpTestServer = new HttpTestServer(moduleUnderTest);
@@ -24,7 +24,7 @@ describe('Unit | Application | Assessment-Results | Routes', () => {
       };
 
       // when
-      const response = await httpTestServer.request('POST', '/api/admin/assessment-results/neutralize-challenge', payload);
+      const response = await httpTestServer.request('POST', '/api/admin/certification/neutralize-challenge', payload);
 
       // then
       expect(response.statusCode).to.equal(403);
@@ -32,7 +32,7 @@ describe('Unit | Application | Assessment-Results | Routes', () => {
 
     it('checks that a valid certification-course id is given', async () => {
       // given
-      sinon.stub(assessmentResultSessionController, 'neutralizeChallenge').returns('ok');
+      sinon.stub(certificationController, 'neutralizeChallenge').returns('ok');
       sinon.stub(securityPreHandlers, 'checkUserHasRolePixMaster');
       securityPreHandlers.checkUserHasRolePixMaster.callsFake((request, h) => h.response(true));
       const httpTestServer = new HttpTestServer(moduleUnderTest);
@@ -46,7 +46,7 @@ describe('Unit | Application | Assessment-Results | Routes', () => {
       };
 
       // when
-      const response = await httpTestServer.request('POST', '/api/admin/assessment-results/neutralize-challenge', payload);
+      const response = await httpTestServer.request('POST', '/api/admin/certification/neutralize-challenge', payload);
 
       // then
       expect(response.statusCode).to.equal(400);
@@ -54,7 +54,7 @@ describe('Unit | Application | Assessment-Results | Routes', () => {
 
     it('checks that a challenge recId is given', async() => {
       // given
-      sinon.stub(assessmentResultSessionController, 'neutralizeChallenge').returns('ok');
+      sinon.stub(certificationController, 'neutralizeChallenge').returns('ok');
       sinon.stub(securityPreHandlers, 'checkUserHasRolePixMaster');
       securityPreHandlers.checkUserHasRolePixMaster.callsFake((request, h) => h.response(true));
       const httpTestServer = new HttpTestServer(moduleUnderTest);
@@ -68,7 +68,7 @@ describe('Unit | Application | Assessment-Results | Routes', () => {
       };
 
       // when
-      const response = await httpTestServer.request('POST', '/api/admin/assessment-results/neutralize-challenge', payload);
+      const response = await httpTestServer.request('POST', '/api/admin/certification/neutralize-challenge', payload);
 
       // then
       expect(response.statusCode).to.equal(400);
